fix: domain correctness, performance, API consistency (BD-1203)
Address 17 medium-severity findings across domain modeling, infrastructure,
API, frontend, and CI:

- Replace raw Guid with ProjectIdentifier/ZoneIdentifier ValueObjects in
  Documentation BC (Fix 16)
- Change CableSpec.CrossSection from int to decimal with guard overloads (Fix 17)
- Accept actual capture time in Photo.TakenAt (Fix 18)
- Validate parent zone exists in Project.AddZone (Fix 19)
- Scope zone name uniqueness to parent zone (Fix 20)
- Allow zero measurement values (rename rule to MeasurementValueMustBeNonNegative) (Fix 21)
- Add Installation update methods with PUT endpoint (Fix 22)
- Add AsNoTracking for read queries via GetByIdReadOnlyAsync (Fix 23)
- Guard against zero PageSize in PagedResult (Fix 24)
- Include full GPS data in PhotoDto (Fix 25)
- Add database indexes on project_id, zone_id, status columns (Fix 27)
- Filter ProjectCountProvider by active status (Fix 28)
- Add OpenAPI metadata to all Documentation endpoints (Fix 31)
- Fix frontend getRootZones to use isNull instead of eq("") (Fix 32)
- Add sync concurrency guard in SyncManager (Fix 33)
- Remove non-null assertion in SyncProvider (Fix 34)
- Pin Trivy action version and enable exit-code failure (Fix 35)

+++ src/backend/Services/Documentation/BauDoku.Documentation.Domain/ValueObjects/CableSpec.cs
@@ -9,11 +9,11 @@ public sealed record CableSpec : ValueObject
     public const int MaxColorLength = 50;
 
     public string CableType { get; }
-    public int? CrossSection { get; }
+    public decimal? CrossSection { get; }
     public string? Color { get; }
     public int? ConductorCount { get; }
 
-    private CableSpec(string cableType, int? crossSection, string? color, int? conductorCount)
+    private CableSpec(string cableType, decimal? crossSection, string? color, int? conductorCount)
     {
         CableType = cableType;
         CrossSection = crossSection;
@@ -21,7 +21,7 @@ public sealed record CableSpec : ValueObject
         ConductorCount = conductorCount;
     }
 
-    public static CableSpec Create(string cableType, int? crossSection = null, string? color = null, int? conductorCount = null)
+    public static CableSpec Create(string cableType, decimal? crossSection = null, string? color = null, int? conductorCount = null)
     {
         Ensure.That(cableType).IsNotNullOrWhiteSpace("Kabeltyp darf nicht leer sein.")
             .MaxLengthIs(MaxCableTypeLength, $"Kabeltyp darf max. {MaxCableTypeLength} Zeichen lang sein.");
